requirements.txt
asgiref==3.8.1
azure-ai-inference==1.0.0b9
azure-core==1.34.0
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
Django==5.2.1
fonttools==4.58.0
idna==3.10
isodate==0.7.2
kiwisolver==1.4.8
matplotlib==3.10.3
narwhals==1.39.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
plotly==5.18.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tenacity==9.1.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
